+++ library/src/main/java/com/cz/android/sample/library/component/message/SampleSystemConsole.java
@@ -3,7 +3,6 @@ package com.cz.android.sample.library.component.message;
 import com.cz.android.sample.library.thread.WorkThread;
 
 import java.io.IOException;
-import java.io.ObjectInputStream;
 import java.io.PipedInputStream;
 import java.io.PipedOutputStream;
 import java.io.PrintStream;
@@ -16,16 +15,10 @@ public class SampleSystemConsole{
     private ReadStreamRunnable readStreamRunnable=null;
 
     public void setup(WorkThread<String> workThread){
-        PipedOutputStream pipedOutputStream=null;
-        PipedInputStream pipedInputStream = new PipedInputStream();
-        try {
-            pipedOutputStream = new PipedOutputStream(pipedInputStream);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        PipedOutputStream pout = new PipedOutputStream();
         // set up System.out
-        System.setOut(new PrintStream(pipedOutputStream, true));
-        readStreamRunnable=new ReadStreamRunnable(workThread,pipedInputStream);
+        System.setOut(new PrintStream(pout, true));
+        readStreamRunnable=new ReadStreamRunnable(workThread,pout);
         executorService.execute(readStreamRunnable);
     }
 
@@ -36,20 +29,20 @@ public class SampleSystemConsole{
     }
 
     private static class ReadStreamRunnable implements Runnable{
-        private final PipedInputStream inputStream;
+        private final PipedOutputStream outputStream;
         private final WorkThread<String> workThread;
         private volatile boolean isRunning=true;
 
-        public ReadStreamRunnable(WorkThread<String> workThread,PipedInputStream inputStream) {
+        public ReadStreamRunnable(WorkThread<String> workThread,PipedOutputStream outputStream) {
             this.workThread=workThread;
-            this.inputStream = inputStream;
+            this.outputStream = outputStream;
         }
 
         public void stop(){
             isRunning=false;
             System.setOut(null);
             try {
-                inputStream.close();
+                outputStream.close();
             } catch (IOException e) {
                 e.printStackTrace();
             }
@@ -58,18 +51,18 @@ public class SampleSystemConsole{
         @Override
         public void run() {
             final byte[] buf = new byte[1024];
-            while (isRunning) {
-                try(ObjectInputStream ois = new ObjectInputStream(inputStream)){
-                    int len = ois.read(buf);
+            try(PipedInputStream is = new PipedInputStream(outputStream)){
+                while (isRunning) {
+                    int len = is.read(buf);
                     if (len == -1) {
                         continue;
                     }
                     String text = new String(buf, 0, len);
                     workThread.post(text);
-                } catch (IOException e){
-                    isRunning=false;
-                    e.printStackTrace();
                 }
+            } catch (IOException e){
+                isRunning=false;
+                e.printStackTrace();
             }
         }
     }